requirements.txt
python-dotenv>=1.0.1
httpx>=0.27.0
openai>=1.0.0
